v2.0.2 后台页面完成书本表格，编辑插入暂未完成

+++ src/main/java/com/bookshop/controller/BookController.java
@@ -29,7 +29,7 @@ public class BookController {
 
     @RequestMapping("/getBook")
     @ResponseBody
-    public responseFromServer getBook(@RequestBody Long bookId){
+    public responseFromServer getBook(@RequestBody Integer bookId){
         return bookService.getBook(bookId);
     }
 
@@ -68,7 +68,7 @@ public class BookController {
 
     @RequestMapping("/deleteBookById")
     @ResponseBody
-    public responseFromServer deleteBookById(@RequestBody Long bookId){
+    public responseFromServer deleteBookById(@RequestBody Integer bookId){
         return bookService.deleteBookById(bookId);
     }
