+++ src/main/java/com/bookshop/service/OrderService.java
@@ -14,8 +14,8 @@ import java.util.Map;
  * @Date: 2020/3/28 2:01
  */
 public interface OrderService {
-    responseFromServer getOrder(Long orderId);
-    responseFromServer getOrdersByUserId(Long userId);
+    responseFromServer getOrder(Integer orderId);
+    responseFromServer getOrdersByUserId(Integer userId);
     responseFromServer insertOrder(Order order);
 
     responseFromServer searchOrdersPage(Map<String,Object> queryMap);
@@ -24,12 +24,12 @@ public interface OrderService {
     responseFromServer insertOrders(List<Order> orders);
     responseFromServer updateOrder(Order order);
     responseFromServer updateOrders(List<Order> orders);
-    responseFromServer deleteOrderById(Long orderId);
+    responseFromServer deleteOrderById(Integer orderId);
     responseFromServer deleteOrder(Order order);
     responseFromServer deleteOrders(List<Order> orders);
     responseFromServer insertOrderItem(OrderItem orderItem);
     responseFromServer updateOrderItem(OrderItem orderItem);
     responseFromServer updateOrderItem(List<OrderItem> orderItems);
     responseFromServer deleteOrderItem(OrderItem orderItem);
-    responseFromServer deleteOrderItemsByOrderId(Long orderId);
+    responseFromServer deleteOrderItemsByOrderId(Integer orderId);
 }
