+++ src/main/java/com/bookshop/service/serviceImpl/CartServiceImplementation.java
@@ -38,7 +38,7 @@ public class CartServiceImplementation {
         if(response.isSuccess()){
             Cart newCart = new Cart();
             newCart.setBooks((List<CartItem>)response.getData());
-            newCart.setUserId((Long)requestMap.get("userId"));
+            newCart.setUserId((Integer)requestMap.get("userId"));
             return responseFromServer.success(newCart);
         }else{
             return responseFromServer.error();
